refactored sampling operators
