+++ src/edu/washington/escience/myria/api/encoding/OperatorEncoding.java
@@ -49,8 +49,7 @@ import edu.washington.escience.myria.operator.SampledDbInsertTemp;
     @Type(name = "RightHashCountingJoin", value = RightHashCountingJoinEncoding.class),
     @Type(name = "RightHashJoin", value = RightHashJoinEncoding.class),
     @Type(name = "SampledDbInsertTemp", value = TempSampledInsertEncoding.class),
-    @Type(name = "SampleWR", value = SampleWREncoding.class),
-    @Type(name = "SampleWoR", value = SampleWoREncoding.class),
+    @Type(name = "Sample", value = SampleEncoding.class),
     @Type(name = "SamplingDistribution", value = SamplingDistributionEncoding.class),
     @Type(name = "SeaFlowScan", value = SeaFlowFileScanEncoding.class),
     @Type(name = "SetGlobal", value = SetGlobalEncoding.class),
