+++ src/edu/washington/escience/myria/api/encoding/SampleEncoding.java
@@ -1,0 +1,18 @@
+package edu.washington.escience.myria.api.encoding;
+
+import edu.washington.escience.myria.api.encoding.QueryConstruct.ConstructArgs;
+import edu.washington.escience.myria.operator.Sample;
+
+public class SampleEncoding extends BinaryOperatorEncoding<Sample> {
+
+  /** Used to make results deterministic. Null if no specified value. */
+  public Long randomSeed;
+
+  @Required
+  public boolean isWithoutReplacement;
+
+  @Override
+  public Sample construct(final ConstructArgs args) {
+    return new Sample(null, null, isWithoutReplacement, randomSeed);
+  }
+}
